Updated to python 3.10, created utils file, moved ValueFinder into own module
