+++ README.md
@@ -2,9 +2,10 @@
 
 ### Requirements
 #### Chrome beta at least v104 and corresponding chromedriver:
-1. Install Chrome Beta from https://www.google.com/chrome/beta/
-2. After installing Chrome Beta, access: "chrome://settings/help" and note the version
-3. Download the corresponding chromedriver_win32.zip from here: https://chromedriver.chromium.org/downloads
+1. Install python 3.10
+2. Install Chrome Beta from https://www.google.com/chrome/beta/
+3. After installing Chrome Beta, access: "chrome://settings/help" and note the version
+4. Download the corresponding chromedriver_win32.zip from here: https://chromedriver.chromium.org/downloads
 
 ### Usage Instructions
 1. Inside your virtual python environment, run pip install -r requirements.txt
