+++ requirements.txt
@@ -1,4 +1,4 @@
-selenium~=4.1.0
-beautifulsoup4~=4.10.0
-requests~=2.27.1
+selenium==4.4.3
+beautifulsoup4==4.11.1
+requests==2.28.1
 XlsxWriter~=3.0.2
