+++ utils.py
@@ -1,0 +1,24 @@
+from datetime import datetime
+
+from selenium import webdriver
+from selenium.common.exceptions import WebDriverException
+from selenium.webdriver.chrome.service import Service
+
+
+def init_time():
+    global current_time
+    current_time = datetime.now()
+
+
+def init_driver(chrome_path, chromedriver_path):
+    options = webdriver.ChromeOptions()
+    options.binary_location = chrome_path
+    options.add_experimental_option("detach", True)
+
+    try:
+        driver = webdriver.Chrome(service=Service(chromedriver_path), options=options)
+    except WebDriverException:
+        print("Wrong paths!")
+        return None
+
+    return driver
